requirements.txt
appdirs
lxml
pybtex
pydantic
pyyaml
